signedness correctness
--HG--
extra : convert_revision : svn%3A16f47a83-c923-0410-902c-c9163ce98875/trunk%401699

+++ src/core/qxtdatastreamsignalserializer.cpp
@@ -85,5 +85,5 @@ QxtAbstractSignalSerializer::DeserializedData QxtDataStreamSignalSerializer::des
 }
 
 bool QxtDataStreamSignalSerializer::canDeserialize(const QByteArray& buffer) const {
-    return qFromLittleEndian<quint32>(reinterpret_cast<const uchar*>(buffer.constData())) >= (buffer.length() - 4);
+    return qFromLittleEndian<quint32>(reinterpret_cast<const uchar*>(buffer.constData())) >= quint32(buffer.length() - 4);
 }
